ghash: Turn size into guint
The size attribute is actually a "capacity" instead of a size in a more
classical sense: It doesn't represent the byte size but the amount of
elements within the hash table.

This is always a power of two, capped at 32 bit. This limit is required
to satisfy the support of converting a GHash into a GArray, which is
capped at G_MAXUINT elements.

Treat the size as capacity and bring it to the same data type as nnodes
and other element counters for better performance.

While at it, turn all variables taking the value of size into a guint
as well.

Closes: #672

+++ glib/tests/hash.c
@@ -1471,7 +1471,7 @@ test_new_similar (void)
 
 struct _GHashTable
 {
-  gsize            size;
+  guint            size;
   gint             mod;
   guint            mask;
   gint             nnodes;
